Домашнє завдання №2 Node
