+++ contlollers/contacts/index.js
@@ -1,0 +1,12 @@
+const getAll = require("./getAll");
+const getById = require("./getById");
+const add = require("./add");
+const updateById = require("./updateById");
+const removeById = require("./removeById");
+module.exports = {
+  getAll,
+  getById,
+  add,
+  updateById,
+  removeById,
+};
